Move keypair from `nano/secure` to `nano/lib`
