+++ nano/lib/keypair.cpp
@@ -1,0 +1,26 @@
+#include <nano/crypto_lib/random_pool.hpp>
+#include <nano/lib/keypair.hpp>
+
+#include <crypto/ed25519-donna/ed25519.h>
+
+// Create a new random keypair
+nano::keypair::keypair ()
+{
+	random_pool::generate_block (prv.bytes.data (), prv.bytes.size ());
+	ed25519_publickey (prv.bytes.data (), pub.bytes.data ());
+}
+
+// Create a keypair given a private key
+nano::keypair::keypair (nano::raw_key && prv_a) :
+	prv (std::move (prv_a))
+{
+	ed25519_publickey (prv.bytes.data (), pub.bytes.data ());
+}
+
+// Create a keypair given a hex string of the private key
+nano::keypair::keypair (std::string const & prv_a)
+{
+	[[maybe_unused]] auto error (prv.decode_hex (prv_a));
+	debug_assert (!error);
+	ed25519_publickey (prv.bytes.data (), pub.bytes.data ());
+}
