+++ nano/secure/common.hpp
@@ -7,6 +7,7 @@
 #include <nano/lib/constants.hpp>
 #include <nano/lib/epochs.hpp>
 #include <nano/lib/fwd.hpp>
+#include <nano/lib/keypair.hpp>
 #include <nano/lib/numbers.hpp>
 #include <nano/lib/object_stream.hpp>
 #include <nano/lib/timer.hpp>
@@ -18,20 +19,6 @@
 
 namespace nano
 {
-/**
- * A key pair. The private key is generated from the random pool, or passed in
- * as a hex string. The public key is derived using ed25519.
- */
-class keypair
-{
-public:
-	keypair ();
-	keypair (std::string const &);
-	keypair (nano::raw_key &&);
-	nano::public_key pub;
-	nano::raw_key prv;
-};
-
 class endpoint_key final
 {
 public:
