+++ nano/secure/fwd.hpp
@@ -6,7 +6,6 @@
 namespace nano
 {
 class account_info;
-class keypair;
 class ledger;
 class ledger_cache;
 class ledger_constants;
